requirements.txt
pandas
google-generativeai
python-dotenv
tabulate
